講座削除APIのプルリクエストを発行

+++ app/Http/Controllers/Api/Instructor/CourseController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api\Instructor;
 
 use App\Model\Course;
+use App\Model\Attendance;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Instructor\CourseDeleteRequest;
 use App\Http\Requests\Instructor\CourseUpdateRequest;
@@ -95,7 +96,7 @@ class CourseController extends Controller
     public function update(CourseUpdateRequest $request)
     {
         $file = $request->file('image');
-
+        
         try {
             $course = Course::FindOrFail($request->course_id);
             $imagePath = $course->image;
@@ -138,9 +139,18 @@ class CourseController extends Controller
     public function delete(CourseDeleteRequest $request)
     {
         $course = Course::findOrFail($request->course_id);
+        if (Attendance::where('course_id', $request->course_id)->exists()) {
+            return response()->json([
+                "result" => false,
+                "message" => "This course has already been taken by students."
+            ]);
+        }
+        if (Storage::exists($course->image)) {
+            Storage::delete($course->image);
+        }
         $course->delete();
         return response()->json([
-
+            "result" => true,
         ]);
     }
 }
